Add gamestore core library files
Introduces new gamestore modules: parsers.lua, player.lua, purchases.lua, senders.lua, and store.lua, implementing store offer parsing, player coin management, purchase processing, and store communication logic. Updates init.lua to integrate these new components for a modular gamestore system.
